backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]
passlib[bcrypt]
email-validator
